# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_patrick_cocobase_internal_models.BatchCreateRequest:
    properties:
      documents:
        items:
          additionalProperties: true
          type: object
        type: array
    required:
    - documents
    type: object
  github_com_patrick_cocobase_internal_models.BatchDeleteRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    required:
    - ids
    type: object
  github_com_patrick_cocobase_internal_models.BatchUpdateRequest:
    properties:
      updates:
        items:
          properties:
            data:
              additionalProperties: true
              type: object
            id:
              type: string
          type: object
        type: array
    required:
    - updates
    type: object
...
  github_com_patrick_cocobase_internal_models.DocumentResponse:
    properties:
      collection_id:
        type: string
      created_at:
        type: string
      data:
        additionalProperties: true
        type: object
      id:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@cocobase.io
    name: Cocobase Support
  description: Backend as a Service with flexible collections and document management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Cocobase API
  version: "1.0"
paths:
...
  /collections/{id}/documents/batch-create:
    post:
      consumes:
      - application/json
      description: Create multiple documents in a single transaction. Optimized for
        bulk inserts.
      parameters:
      - description: Collection ID or Name
        in: path
        name: id
        required: true
        type: string
      - description: Documents to create
        in: body
        name: documents
        required: true
        schema:
          $ref: '#/definitions/github_com_patrick_cocobase_internal_models.BatchCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_patrick_cocobase_internal_models.DocumentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Batch create documents
      tags:
      - Batch Operations
  /collections/{id}/documents/batch-delete:
    post:
      consumes:
      - application/json
      description: Delete multiple documents at once. Single query for optimal performance.
      parameters:
      - description: Collection ID or Name
        in: path
        name: id
        required: true
        type: string
      - description: Document IDs to delete
        in: body
        name: deletes
        required: true
        schema:
          $ref: '#/definitions/github_com_patrick_cocobase_internal_models.BatchDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Batch delete documents
      tags:
      - Batch Operations
  /collections/{id}/documents/batch-update:
    post:
      consumes:
      - application/json
      description: Update multiple documents at once. Batch operation reduces queries.
      parameters:
      - description: Collection ID or Name
        in: path
        name: id
        required: true
        type: string
      - description: Document updates
        in: body
        name: updates
        required: true
        schema:
          $ref: '#/definitions/github_com_patrick_cocobase_internal_models.BatchUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Batch update documents
      tags:
      - Batch Operations
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and API key.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"